// generated/graphql.ts
export type Asset = Node & {
	__typename?: 'Asset';
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	fileSize: Scalars['Int']['output'];
	focalPoint?: Maybe<Coordinate>;
	height: Scalars['Int']['output'];
	id: Scalars['ID']['output'];
	mimeType: Scalars['String']['output'];
	name: Scalars['String']['output'];
	preview: Scalars['String']['output'];
	source: Scalars['String']['output'];
	tags: Array<Tag>;
	type: AssetType;
	updatedAt: Scalars['DateTime']['output'];
	width: Scalars['Int']['output'];
};
// ...
export type CurrencyCode = (typeof CurrencyCode)[keyof typeof CurrencyCode];
// ...
export type Customer = Node & {
	__typename?: 'Customer';
	addresses?: Maybe<Array<Address>>;
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	emailAddress: Scalars['String']['output'];
	firstName: Scalars['String']['output'];
	id: Scalars['ID']['output'];
	lastName: Scalars['String']['output'];
	orders: OrderList;
	phoneNumber?: Maybe<Scalars['String']['output']>;
	title?: Maybe<Scalars['String']['output']>;
	updatedAt: Scalars['DateTime']['output'];
	user?: Maybe<User>;
};
// ...
export const ErrorCode = {
	AlreadyLoggedInError: 'ALREADY_LOGGED_IN_ERROR',
	CouponCodeExpiredError: 'COUPON_CODE_EXPIRED_ERROR',
	CouponCodeInvalidError: 'COUPON_CODE_INVALID_ERROR',
	CouponCodeLimitError: 'COUPON_CODE_LIMIT_ERROR',
	EmailAddressConflictError: 'EMAIL_ADDRESS_CONFLICT_ERROR',
	GuestCheckoutError: 'GUEST_CHECKOUT_ERROR',
	IdentifierChangeTokenExpiredError: 'IDENTIFIER_CHANGE_TOKEN_EXPIRED_ERROR',
	IdentifierChangeTokenInvalidError: 'IDENTIFIER_CHANGE_TOKEN_INVALID_ERROR',
	IneligiblePaymentMethodError: 'INELIGIBLE_PAYMENT_METHOD_ERROR',
	IneligibleShippingMethodError: 'INELIGIBLE_SHIPPING_METHOD_ERROR',
	InsufficientStockError: 'INSUFFICIENT_STOCK_ERROR',
	InvalidCredentialsError: 'INVALID_CREDENTIALS_ERROR',
	MissingPasswordError: 'MISSING_PASSWORD_ERROR',
	NativeAuthStrategyError: 'NATIVE_AUTH_STRATEGY_ERROR',
	NegativeQuantityError: 'NEGATIVE_QUANTITY_ERROR',
	NotVerifiedError: 'NOT_VERIFIED_ERROR',
	NoActiveOrderError: 'NO_ACTIVE_ORDER_ERROR',
	OrderInterceptorError: 'ORDER_INTERCEPTOR_ERROR',
	OrderLimitError: 'ORDER_LIMIT_ERROR',
	OrderModificationError: 'ORDER_MODIFICATION_ERROR',
	OrderPaymentStateError: 'ORDER_PAYMENT_STATE_ERROR',
	OrderStateTransitionError: 'ORDER_STATE_TRANSITION_ERROR',
	PasswordAlreadySetError: 'PASSWORD_ALREADY_SET_ERROR',
	PasswordResetTokenExpiredError: 'PASSWORD_RESET_TOKEN_EXPIRED_ERROR',
	PasswordResetTokenInvalidError: 'PASSWORD_RESET_TOKEN_INVALID_ERROR',
	PasswordValidationError: 'PASSWORD_VALIDATION_ERROR',
	PaymentDeclinedError: 'PAYMENT_DECLINED_ERROR',
	PaymentFailedError: 'PAYMENT_FAILED_ERROR',
	UnknownError: 'UNKNOWN_ERROR',
	VerificationTokenExpiredError: 'VERIFICATION_TOKEN_EXPIRED_ERROR',
	VerificationTokenInvalidError: 'VERIFICATION_TOKEN_INVALID_ERROR',
} as const;

export type ErrorCode = (typeof ErrorCode)[keyof typeof ErrorCode];
// ...
export type IneligiblePaymentMethodError = ErrorResult & {
	__typename?: 'IneligiblePaymentMethodError';
	eligibilityCheckerMessage?: Maybe<Scalars['String']['output']>;
	errorCode: ErrorCode;
	message: Scalars['String']['output'];
};
// ...
export type NoActiveOrderError = ErrorResult & {
	__typename?: 'NoActiveOrderError';
	errorCode: ErrorCode;
	message: Scalars['String']['output'];
};

export type Node = {
	id: Scalars['ID']['output'];
};
// ...
export type Order = Node & {
	__typename?: 'Order';
	/** An order is active as long as the payment process has not been completed */
	active: Scalars['Boolean']['output'];
	billingAddress?: Maybe<OrderAddress>;
	/** A unique code for the Order */
	code: Scalars['String']['output'];
	/** An array of all coupon codes applied to the Order */
	couponCodes: Array<Scalars['String']['output']>;
	createdAt: Scalars['DateTime']['output'];
	currencyCode: CurrencyCode;
	customFields?: Maybe<Scalars['JSON']['output']>;
	customer?: Maybe<Customer>;
	discounts: Array<Discount>;
	fulfillments?: Maybe<Array<Fulfillment>>;
	history: HistoryEntryList;
	id: Scalars['ID']['output'];
	lines: Array<OrderLine>;
	/**
	 * The date & time that the Order was placed, i.e. the Customer
	 * completed the checkout and the Order is no longer "active"
	 */
	orderPlacedAt?: Maybe<Scalars['DateTime']['output']>;
	payments?: Maybe<Array<Payment>>;
	/** Promotions applied to the order. Only gets populated after the payment process has completed. */
	promotions: Array<Promotion>;
	shipping: Scalars['Money']['output'];
	shippingAddress?: Maybe<OrderAddress>;
	shippingLines: Array<ShippingLine>;
	shippingWithTax: Scalars['Money']['output'];
	state: Scalars['String']['output'];
	/**
	 * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
	 * discounts which have been prorated (proportionally distributed) amongst the items of each OrderLine.
	 * To get a total of all OrderLines which does not account for prorated discounts, use the
	 * sum of `OrderLine.discountedLinePrice` values.
	 */
	subTotal: Scalars['Money']['output'];
	/** Same as subTotal, but inclusive of tax */
	subTotalWithTax: Scalars['Money']['output'];
	/**
	 * Surcharges are arbitrary modifications to the Order total which are neither
	 * ProductVariants nor discounts resulting from applied Promotions. For example,
	 * one-off discounts based on customer interaction, or surcharges based on payment
	 * methods.
	 */
	surcharges: Array<Surcharge>;
	/** A summary of the taxes being applied to this Order */
	taxSummary: Array<OrderTaxSummary>;
	/** Equal to subTotal plus shipping */
	total: Scalars['Money']['output'];
	totalQuantity: Scalars['Int']['output'];
	/** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
	totalWithTax: Scalars['Money']['output'];
	type: OrderType;
	updatedAt: Scalars['DateTime']['output'];
};

export type OrderHistoryArgs = {
	options?: InputMaybe<HistoryEntryListOptions>;
};

export type OrderAddress = {
	__typename?: 'OrderAddress';
	city?: Maybe<Scalars['String']['output']>;
	company?: Maybe<Scalars['String']['output']>;
	country?: Maybe<Scalars['String']['output']>;
	countryCode?: Maybe<Scalars['String']['output']>;
	customFields?: Maybe<Scalars['JSON']['output']>;
	fullName?: Maybe<Scalars['String']['output']>;
	phoneNumber?: Maybe<Scalars['String']['output']>;
	postalCode?: Maybe<Scalars['String']['output']>;
	province?: Maybe<Scalars['String']['output']>;
	streetLine1?: Maybe<Scalars['String']['output']>;
	streetLine2?: Maybe<Scalars['String']['output']>;
};
// ...
export type OrderTaxSummary = {
	__typename?: 'OrderTaxSummary';
	/** A description of this tax */
	description: Scalars['String']['output'];
	/** The total net price of OrderLines to which this taxRate applies */
	taxBase: Scalars['Money']['output'];
	/** The taxRate as a percentage */
	taxRate: Scalars['Float']['output'];
	/** The total tax being applied to the Order at this taxRate */
	taxTotal: Scalars['Money']['output'];
};
// ...
export type PaymentInput = {
	/**
	 * This field should contain arbitrary data passed to the specified PaymentMethodHandler's `createPayment()` method
	 * as the "metadata" argument. For example, it could contain an ID for the payment and other
	 * data generated by the payment provider.
	 */
	metadata: Scalars['JSON']['input'];
	/** This field should correspond to the `code` property of a PaymentMethod. */
	method: Scalars['String']['input'];
};
// ...
export type Product = Node & {
	__typename?: 'Product';
	assets: Array<Asset>;
	collections: Array<Collection>;
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	description: Scalars['String']['output'];
	enabled: Scalars['Boolean']['output'];
	facetValues: Array<FacetValue>;
	featuredAsset?: Maybe<Asset>;
	id: Scalars['ID']['output'];
	languageCode: LanguageCode;
	name: Scalars['String']['output'];
	optionGroups: Array<ProductOptionGroup>;
	slug: Scalars['String']['output'];
	translations: Array<ProductTranslation>;
	updatedAt: Scalars['DateTime']['output'];
	/** Returns a paginated, sortable, filterable list of ProductVariants */
	variantList: ProductVariantList;
	/** Returns all ProductVariants */
	variants: Array<ProductVariant>;
};
// ...
export type ProductVariant = Node & {
	__typename?: 'ProductVariant';
	assets: Array<Asset>;
	createdAt: Scalars['DateTime']['output'];
	currencyCode: CurrencyCode;
	customFields?: Maybe<Scalars['JSON']['output']>;
	facetValues: Array<FacetValue>;
	featuredAsset?: Maybe<Asset>;
	id: Scalars['ID']['output'];
	languageCode: LanguageCode;
	name: Scalars['String']['output'];
	options: Array<ProductOption>;
	price: Scalars['Money']['output'];
	priceWithTax: Scalars['Money']['output'];
	product: Product;
	productId: Scalars['ID']['output'];
	sku: Scalars['String']['output'];
	stockLevel: Scalars['String']['output'];
	taxCategory: TaxCategory;
	taxRateApplied: TaxRate;
	translations: Array<ProductVariantTranslation>;
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type ShippingLine = {
	__typename?: 'ShippingLine';
	customFields?: Maybe<Scalars['JSON']['output']>;
	discountedPrice: Scalars['Money']['output'];
	discountedPriceWithTax: Scalars['Money']['output'];
	discounts: Array<Discount>;
	id: Scalars['ID']['output'];
	price: Scalars['Money']['output'];
	priceWithTax: Scalars['Money']['output'];
	shippingMethod: ShippingMethod;
};

export type ShippingMethod = Node & {
	__typename?: 'ShippingMethod';
	calculator: ConfigurableOperation;
	checker: ConfigurableOperation;
	code: Scalars['String']['output'];
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	description: Scalars['String']['output'];
	fulfillmentHandlerCode: Scalars['String']['output'];
	id: Scalars['ID']['output'];
	languageCode: LanguageCode;
	name: Scalars['String']['output'];
	translations: Array<ShippingMethodTranslation>;
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type AddPaymentToOrderMutation = {
	__typename?: 'Mutation';
	addPaymentToOrder:
		| { __typename?: 'IneligiblePaymentMethodError'; errorCode: ErrorCode; message: string }
		| { __typename?: 'NoActiveOrderError'; errorCode: ErrorCode; message: string }
		| {
				__typename: 'Order';
				id: string;
				code: string;
				active: boolean;
				createdAt: any;
				state: string;
				currencyCode: CurrencyCode;
				totalQuantity: number;
				subTotal: any;
				subTotalWithTax: any;
				shippingWithTax: any;
				totalWithTax: any;
				taxSummary: Array<{
					__typename?: 'OrderTaxSummary';
					description: string;
					taxRate: number;
					taxTotal: any;
				}>;
				customer?: {
					__typename?: 'Customer';
					id: string;
					firstName: string;
					lastName: string;
					emailAddress: string;
				} | null;
				shippingAddress?: {
					__typename?: 'OrderAddress';
					fullName?: string | null;
					streetLine1?: string | null;
					streetLine2?: string | null;
					company?: string | null;
					city?: string | null;
					province?: string | null;
					postalCode?: string | null;
					countryCode?: string | null;
					phoneNumber?: string | null;
				} | null;
				shippingLines: Array<{
					__typename?: 'ShippingLine';
					priceWithTax: any;
					shippingMethod: { __typename?: 'ShippingMethod'; id: string; name: string };
				}>;
				lines: Array<{
					__typename?: 'OrderLine';
					id: string;
					unitPriceWithTax: any;
					linePriceWithTax: any;
					quantity: number;
					featuredAsset?: { __typename?: 'Asset'; id: string; preview: string } | null;
					productVariant: {
						__typename?: 'ProductVariant';
						id: string;
						name: string;
						price: any;
						product: { __typename?: 'Product'; id: string; slug: string };
					};
				}>;
		  }
		| { __typename?: 'OrderPaymentStateError'; errorCode: ErrorCode; message: string }
		| { __typename?: 'OrderStateTransitionError'; errorCode: ErrorCode; message: string }
		| { __typename?: 'PaymentDeclinedError'; errorCode: ErrorCode; message: string }
		| { __typename?: 'PaymentFailedError'; errorCode: ErrorCode; message: string };
};